% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{apply_imputation}
\alias{apply_imputation}
\title{apply_imputation}
\usage{
apply_imputation(imputeData, omicsData)
}
\arguments{
\item{imputeData}{object of class imputeData that contains the imputed e_data
determined using the function 'imputation'}

\item{omicsData}{an object of the class 'pepData', 'proData', 'metabData',
'lipidData', or 'nmrData', usually created by \code{\link{as.pepData}},
\code{\link{as.proData}}, \code{\link{as.metabData}},
\code{\link{as.lipidData}}, or \code{\link{as.nmrData}}, respectively.}
}
\value{
object of pmartR object containing imputated e_data
}
\description{
This function returns a pmart object that has undergone imputation
}
\examples{
library(malbacR)
data("pmart_amide")
impFilt <- imputation(pmart_amide)
imputed_data <- apply_imputation(impFilt,pmart_amide)

}
\author{
Evan Martin, Damon Leach
}
