% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiger_filtering.R
\name{apply_tigerFilt}
\alias{apply_tigerFilt}
\title{Apply a S3 tiger filter object to a pmartR S3 object}
\usage{
apply_tigerFilt(filter_object, omicsData)
}
\arguments{
\item{filter_object}{an object of the class 'tigerFilt' created by
\code{tiger_filter}}

\item{omicsData}{an object of the class \code{pepData}, \code{proData},
\code{lipidData}, \code{metabData}, or \code{nmrData} usually created by
\code{\link{as.pepData}}, \code{\link{as.proData}},
\code{\link{as.lipidData}}, \code{\link{as.metabData}}, or \code{\link{as.nmrData}}, respectively.}
}
\value{
An object of the class \code{pepData}, \code{proData},
  \code{lipidData}, \code{metabData}, or \code{nmrData} with specified cname_ids,
  edata_cnames, and emeta_cnames filtered out of the appropriate datasets.
}
\description{
This function takes a filter object of class 'tigerFilt' and 
  applies the filter to a dataset of \code{pepData}, \code{proData}, 
  \code{lipidData}, \code{metabData}, or \code{nmrData}.
}
\examples{
dontrun{
library(pmartRdata)
data("pmart_amide")
to_filter <- tiger_filter(omicsData = pmart_amide,sampletype_cname = "group", test_val = "QC")
amideFilt <- apply_tigerFilt(filter_object = to_filter, omicsData = pep_object)
}

}
\author{
Lisa Bramer, Kelly Stratton,Damon Leach
}
