% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_combat.R
\name{bc_combat}
\alias{bc_combat}
\title{ComBat adjusted object}
\usage{
bc_combat(omicsData, use_groups = FALSE)
}
\arguments{
\item{omicsData}{an object of the class 'pepData', 'proData', 'metabData',
'lipidData', or 'nmrData', usually created by \code{\link{as.pepData}},
\code{\link{as.proData}}, \code{\link{as.metabData}},
\code{\link{as.lipidData}}, or \code{\link{as.nmrData}}, respectively.}

\item{use_groups}{logical argument indicating if the group designation should
be applied to the data for batch correction. Defaults to FALSE. If TRUE,
ComBat runs with both a batch effect and a group designation}
}
\value{
Object of same class as omicsData that has been undergone
  ComBat batch effect correction
}
\description{
This function returns a pmart object that has been undergone ComBat
batch effect correction
}
\examples{
\dontrun{
library(danceR)
data("pmart_amide")
pmart_amide <- edata_transform(pmart_amide,"log2")
pmart_amide <- group_designation(pmart_amide,main_effects = "group",batch_id = "batch")
pmart_amide <- normalize_global(pmart_amide,subset_fn = "all",norm_fn = "median",
                               apply_norm = TRUE,backtransform = TRUE)
amide_combat <- bc_combat(omicsData = pmart_amide, use_groups = FALSE)
}

}
\author{
Damon Leach
}
