% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_range.R
\name{bc_range}
\alias{bc_range}
\title{Range Scaling adjusted object}
\usage{
bc_range(omicsData)
}
\arguments{
\item{omicsData}{an object of the class 'pepData', 'proData', 'metabData',
'lipidData', or 'nmrData', usually created by \code{\link{as.pepData}},
\code{\link{as.proData}}, \code{\link{as.metabData}},
\code{\link{as.lipidData}}, or \code{\link{as.nmrData}}, respectively.}
}
\value{
Object of same class as omicsData that has been undergone
  Range Scaling Normalization as batch effect correction
}
\description{
This function returns a pmart object that has been undergone Range Scaling
normalization
}
\examples{
library(malbacR)
library(pmartR)
data("pmart_amide")
pmart_amide <- edata_transform(pmart_amide,"log2")
pmart_amide <- group_designation(pmart_amide,main_effects = "group",batch_id = "batch")
pmart_amide <- normalize_global(pmart_amide,subset_fn = "all",norm_fn = "median",
                               apply_norm = TRUE,backtransform = TRUE)
amide_range <- bc_range(omicsData = pmart_amide)

}
\author{
Damon Leach
}
