% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_tiger.R
\name{bc_tiger}
\alias{bc_tiger}
\title{TIGER batch correction}
\usage{
bc_tiger(
  omicsData,
  sampletype_cname,
  test_val,
  position_cname = NULL,
  injection_cname = NULL
)
}
\arguments{
\item{omicsData}{an object of the class 'pepData', 'proData', 'metabData',
'lipidData', or 'nmrData', usually created by \code{\link{as.pepData}},
\code{\link{as.proData}}, \code{\link{as.metabData}},
\code{\link{as.lipidData}}, or \code{\link{as.nmrData}}, respectively.}

\item{sampletype_cname}{character string giving the name of the column in omicsData$f_data
that contains the sample type information (such as quality control samples)}

\item{test_val}{character string giving the name of the value within the column sampletype_cname to be used
as the testing value for TIGER}

\item{position_cname}{character string giving the name of the column in omicsData$f_data that
contains the well position, default is NULL}

\item{injection_cname}{character string giving the name of the column in omicsData$f_data that
contains the injection order, default is NULL}
}
\value{
Object of same class as omicsData that has been undergone
  TIGER normalization
}
\description{
This function returns a pmart object that has been undergone TIGER
 batch effect correction
}
\examples{
\dontrun{
library(danceR)
data("pmart_amide")
pmart_amide <- edata_transform(pmart_amide,"log2")
pmart_amide <- group_designation(pmart_amide,main_effects = "group",batch_id = "batch")
pmart_amide <- normalize_global(pmart_amide,subset_fn = "all",norm_fn = "median",
                               apply_norm = TRUE,backtransform = TRUE)
tigerFilt <- tiger_filter(pmart_amide,sampletype_cname = "group",test_val = "QC")
pmart_amideFilt <- apply_tigerFilt(tigerFilt,pmart_amide)
amide_tiger <- bc_tiger(omicsData = pmart_amideFilt,sampletype_cname = "group",
                        test_val = "QC",injection_cname = "Injection_order")
}

}
\author{
Damon Leach
}
