% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_waveica.R
\name{bc_waveica}
\alias{bc_waveica}
\title{WaveICA batch correction}
\usage{
bc_waveica(
  omicsData,
  batch_cname = NULL,
  injection_cname = NULL,
  version = "WaveICA",
  alpha = 0,
  cutoff_injection = 0.1,
  K = 20,
  cutoff_batch = 0.05,
  cutoff_group = 0.05
)
}
\arguments{
\item{omicsData}{an object of the class 'pepData', 'proData', 'metabData',
'lipidData', or 'nmrData', usually created by \code{\link{as.pepData}},
\code{\link{as.proData}}, \code{\link{as.metabData}},
\code{\link{as.lipidData}}, or \code{\link{as.nmrData}}, respectively.}

\item{injection_cname}{character string giving the name of the column in omicsData$f_data
that contains the run order of the samples (to be used for WaveICA2.0)}

\item{version}{character string that specifies whether the user wants to run "WaveICA"
or "WaveICA2.0" (default is "WaveICA")}

\item{alpha}{tradeoff value between the independence of samples and those of variables in ICA,
and should be between 0 and 1, defaults to 0}

\item{cutoff_injection}{threshold of the variation explained by the injection order for independent components,
should be between 0 and 1, defaults to 0.1 (to be used for WaveICA2.0)}

\item{K}{maximal component that ICA decomposes, defaults to 20 (though recommendation for K = 10
when using "WaveICA2.0")}

\item{cutoff_batch}{threshold of the variation explained by the batch for independent components,
should be between 0 and 1, defaults to 0.05 (to be used for WaveICA)}

\item{cutoff_group}{threshold of the variation explained by the group for independent components,
should be between 0 and 1, defaults to 0.05 (to be used for WaveICA)}

\item{batach_cname}{character string giving the name of the colum in omicsData$f_data
that contains the batch information of the samples (to be used for WaveICA)}
}
\value{
Object of same class as omicsData that has been undergone
  WaveICA normalization
}
\description{
This function returns a pmart object that has been undergone either WaveICA or WaveICA_2.0
 batch effect correction
}
\examples{
library(malbacR)
library(pmartR)
data("pmart_amide")
pmart_amide <- edata_transform(pmart_amide,"log2")
pmart_amide <- group_designation(pmart_amide,main_effects = "group",batch_id = "batch")
impObj <- imputation(omicsData = pmart_amide)
amide_imp <- apply_imputation(imputeData = impObj, omicsData = pmart_amide)
amide_wave <- bc_waveica(omicsData = amide_imp, injection_cname = "Injection_order",
                         version = "WaveICA2.0",alpha = 0, cutoff_injection = 0.1, K = 10)

}
\author{
Damon Leach
}
