% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tiger.R
\name{compute_RSD}
\alias{compute_RSD}
\title{compute_RSD}
\usage{
compute_RSD(input_data)
}
\arguments{
\item{input_data}{description}
}
\value{
return tiger batch corrected results
}
\description{
This function computes RSD
}
