% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fns.R
\name{get_batch_method}
\alias{get_batch_method}
\alias{get_data_batch}
\title{Fetch the Logical Vector of if data has been batch corrected}
\usage{
get_data_batch(omicsData)
}
\arguments{
\item{omicsData}{An object of class pepData, proData, metabData, lipidData,
or nmrData.}
}
\value{
A logical value specifying if batch correction has been applied or not (TRUE FALSE if no batch correction applied)
}
\description{
Retrieves the TRUE/FALSE statement of if data has been batch corrected
}
