% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{imputation}
\alias{imputation}
\title{imputation}
\usage{
imputation(omicsData)
}
\arguments{
\item{omicsData}{an object of the class 'pepData', 'proData', 'metabData',
'lipidData', or 'nmrData', usually created by \code{\link{as.pepData}},
\code{\link{as.proData}}, \code{\link{as.metabData}},
\code{\link{as.lipidData}}, or \code{\link{as.nmrData}}, respectively.}
}
\value{
object of class imputeData and dataframe with the imputated e_data
}
\description{
This function returns a pmart filter that has imputed missing values by "em" method
}
\examples{
\dontrun{
library(pmartRdata)
data("mini_udn")
molfilt <- pmartR::molecule_filter(mini_udn)
mini_udn <- pmartR::applyFilt(molfilt, mini_udn)
imp <- imputation(mini_udn)
}

}
\author{
Evan Martin, Damon Leach
}
