% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tiger.R
\name{Internal.boxplot.stats}
\alias{Internal.boxplot.stats}
\title{Internal.boxplot.stats}
\usage{
Internal.boxplot.stats(x, coef = 1.5, do.conf = TRUE, do.out = TRUE)
}
\value{
return boxplot statistics
}
\description{
This function is an internal function run boxplots
}
