% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tiger.R
\name{Internal.compute_errorRatio}
\alias{Internal.compute_errorRatio}
\title{Internal.compute_errorRatio}
\usage{
Internal.compute_errorRatio(rawVal, sampleType, targetVal)
}
\value{
return computed error ratio
}
\description{
This function is an internal function to compute error ratio
}
