% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tiger.R
\name{Internal.compute_targetVal}
\alias{Internal.compute_targetVal}
\title{Internal.compute_targetVal}
\usage{
Internal.compute_targetVal(
  input_col,
  targetVal_method = c("median", "mean"),
  targetVal_removeOutlier = TRUE
)
}
\value{
return computed target values
}
\description{
This function is an internal function to compute target values
}
