% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tiger.R
\name{Internal.createFolds}
\alias{Internal.createFolds}
\title{Internal.createFolds}
\usage{
Internal.createFolds(y, k = 10, list = TRUE, returnTrain = FALSE)
}
\value{
return created folds
}
\description{
This function is an internal function to create folds
}
