% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tiger.R
\name{Internal.remove_NA}
\alias{Internal.remove_NA}
\title{Internal.remove_NA}
\usage{
Internal.remove_NA(
  input_data_num,
  data_label = NULL,
  replace_with_zero = FALSE
)
}
\arguments{
\item{input_data_num}{description}

\item{data_label}{description}

\item{replace_with_zero}{description}
}
\value{
return tiger batch corrected results
}
\description{
This function is an internal function to remove NA values
}
