% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pmart_amideFilt}
\alias{pmart_amideFilt}
\title{Filtered pmartR-friendly "Amide" Data}
\format{
A metabData object (see
  \code{\link[pmartR]{as.metabData}} for details)
\describe{
  \item{e_data}{a \eqn{500 \times 643} data frame of expression data.
  Each row corresponds to data for each metabolite}
  \item{f_data}{a data frame with \eqn{642 \times 4} data frame of feature data.
  Each row corresponds to a sample with a column giving the unique sample identifiers found in e_data
  column names, the injection order, the group information, and the batch information.}
}
}
\description{
These data originates from the pmart_amide dataset. However, all necessary
data transformations and filters have been applied to the data such that
all batch correction methods that in theory can be applied to this dataset can
without any additional alterations.
}
