% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pmart_mix}
\alias{pmart_mix}
\title{pmartR-friendly "mix" Data}
\format{
A metabData object (see
  \code{\link[pmartR]{as.metabData}} for details)
\describe{
  \item{e_data}{a \eqn{46 \times 42} data frame of expression data.
  Each row corresponds to data for each metabolite}
  \item{f_data}{a data frame with \eqn{42 \times 2} data frame of feature data.
  Each row corresponds to a sample with a column giving the unique sample identifiers found in e_data
  column names and for the batch information}
  \item{e_meta}{a data frame with \eqn{46 \times 7} data frame of meta information.
  Each row corresponds to meta information for a metabolite with a column giving the metabolite, marking
  information, internal standards information, synonyms, RI value, query, and known entries}
}
}
\description{
These data contains information for 46 features and 42 samples. These data
originated from the R package "crmn" (see \code{\link[crmn]{mix}} for 
details).
}
