% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pmart_mixFilt}
\alias{pmart_mixFilt}
\title{Filtered pmartR-friendly "mix" Data}
\format{
A metabData object (see
  \code{\link[pmartR]{as.metabData}} for details)
\describe{
  \item{e_data}{a \eqn{46 \times 42} data frame of expression data.
  Each row corresponds to data for each metabolite}
  \item{f_data}{a data frame with \eqn{42 \times 2} data frame of feature data.
  Each row corresponds to a sample with a column giving the unique sample identifiers found in e_data
  column names and for the batch information}
  \item{e_meta}{a data frame with \eqn{46 \times 7} data frame of meta information.
  Each row corresponds to meta information for a metabolite with a column giving the metabolite, marking
  information, internal standards information, synonyms, RI value, query, and known entries}
}
}
\description{
These data originates from the pmart_mix dataset. However, all necessary
data transformations and filters have been applied to the data such that
all batch correction methods that in theory can be applied to this dataset can
without any additional alterations.
}
