% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tiger.R
\name{run_TIGER_internal}
\alias{run_TIGER_internal}
\title{run_TIGER_internal}
\usage{
run_TIGER_internal(
  test_samples,
  train_samples,
  col_sampleID,
  col_sampleType,
  col_batchID,
  col_order = NULL,
  col_position = NULL,
  targetVal_external = NULL,
  targetVal_method = c("mean", "median"),
  targetVal_batchWise = FALSE,
  targetVal_removeOutlier = !targetVal_batchWise,
  selectVar_external = NULL,
  selectVar_corType = c("cor", "pcor"),
  selectVar_corMethod = c("pearson", "spearman"),
  selectVar_minNum = 5,
  selectVar_maxNum = 10,
  selectVar_batchWise = FALSE,
  mtry_percent = seq(0.2, 0.8, 0.2),
  nodesize_percent = seq(0.2, 0.8, 0.2),
  ...,
  parallel.cores = 2
)
}
\arguments{
\item{test_samples}{description}

\item{train_samples}{description}

\item{col_sampleID}{description}

\item{col_sampleType}{description}

\item{col_batchID}{description}

\item{col_order}{description}

\item{col_position}{description}
}
\value{
return tiger batch corrected results
}
\description{
This function is an internal function to run ensembl learning
}
