% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_method.R
\name{scale_method}
\alias{scale_method}
\title{Perform scale method on data}
\usage{
scale_method(
  data,
  methods = c("auto", "range", "pareto", "vast", "level", "power")
)
}
\arguments{
\item{data}{a data matrix or data frame object where each row is a biomolecule
and each column each sample or replicate}

\item{methods}{a string character specifying the pre-treatment method and can be 
"auto", "range", "pareto", "vast", "level", or "power"}
}
\value{
a data matrix or data frame object (depending on what was the input) that
has undergone pre-treatment method
}
\description{
This function is a modified version of the function ScalingMethods from the 
package DiffCorr, but it allows for missing values for Power and Range scaling
}
\examples{
library(malbacR)
data(pmart_amide)
malbacR:::scale_method(pmart_amide$e_data[,-1])

}
\author{
Damon Leach
}
