% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tiger.R
\name{select_variable}
\alias{select_variable}
\title{select_variable}
\usage{
select_variable(
  train_num,
  test_num = NULL,
  train_batchID = NULL,
  test_batchID = NULL,
  selectVar_corType = c("cor", "pcor"),
  selectVar_corMethod = c("spearman", "pearson"),
  selectVar_minNum = 5,
  selectVar_maxNum = 10,
  selectVar_batchWise = FALSE,
  coerce_numeric = FALSE
)
}
\arguments{
\item{train_num}{description}

\item{test_num}{description}

\item{train_batchID}{description}

\item{test_batchID}{description}

\item{selectVar_corType}{description}

\item{selectVar_corMethod}{description}

\item{selectVar_maxNum}{description}

\item{selectVar_batchWise}{description}

\item{coerce_numeric}{description}
}
\value{
return tiger batch corrected results
}
\description{
This function is an internal function run boxplots
}
