% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiger_filtering.R
\name{tiger_filter}
\alias{tiger_filter}
\title{TIGER filtering}
\usage{
tiger_filter(omicsData, sampletype_cname, test_val)
}
\arguments{
\item{omicsData}{an object of the class 'pepData', 'proData', 'metabData',
'lipidData', or 'nmrData', usually created by \code{\link{as.pepData}},
\code{\link{as.proData}}, \code{\link{as.metabData}},
\code{\link{as.lipidData}}, or \code{\link{as.nmrData}}, respectively.}

\item{sampletype_cname}{character string giving the name of the column in omicsData$f_data
that contains the sample type information (such as quality control samples)}

\item{test_val}{character string giving the name of the value within the column sampletype_cname to be used
as the testing value for TIGER}
}
\value{
Object of tigerFilt class
}
\description{
This function returns a object of class 'tigerFilt' that has been filtered to ensure 
TIGER batch correction can run successfully
}
\examples{
\dontrun{
library(pmartRdata)
data(pmart_amide)
tigerFilt <- tiger_filter(omicsData = pmart_amide,sampletype_cname = "group", test_val = "QC")
pmart_amide_filt <- pmartR::applyFilt(filter_object = tigerFilt,omicsData = pmart_amide_filt)
amide_tiger <- bc_tiger(omicsData = pmart_amide_filt, sampletype_cname = "group", test_val = "QC")
}

}
\author{
Damon Leach
}
