% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_nomis.R
\name{bc_nomis}
\alias{bc_nomis}
\title{NOMIS normalized adjusted object}
\usage{
bc_nomis(omicsData, is_cname, is_val, num_pc = 2)
}
\arguments{
\item{omicsData}{an object of the class 'pepData', 'proData', 'metabData',
'lipidData', or 'nmrData', usually created by \code{\link{as.pepData}},
\code{\link{as.proData}}, \code{\link{as.metabData}},
\code{\link{as.lipidData}}, or \code{\link{as.nmrData}}, respectively.}

\item{is_cname}{character string giving the name of the column in omicsData$e_meta
that contains the factor variable indicating whether the molecule is internal
standards or not}

\item{is_val}{value giving the name/value in the is_cname column that indicates
that the sample is an Internal Standards sample}

\item{num_pc}{numeric variable providing the number of principal components for NOMIS
to consider when running the batch correction method, defaults to 2}
}
\value{
Object of same class as omicsData that has been undergone
  NOMIS normalization
}
\description{
This function returns a pmart object that has been undergone NOMIS
normalization regarding batch effect correction
}
\examples{
library(malbacR)
library(pmartR)
data("pmart_mix")
pmart_mix <- edata_transform(pmart_mix,"log2")
pmart_mix <- group_designation(pmart_mix,main_effects = "BatchNum",batch_id = "BatchNum")
pmart_mix <- normalize_global(pmart_mix,subset_fn = "all",norm_fn = "median",
                              apply_norm = TRUE,backtransform = TRUE)
mix_nomis <- bc_nomis(omicsData = pmart_mix, is_cname = "tag", is_val = "IS", num_pc = 2)

}
\author{
Damon Leach
}
