% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_waveica.R
\name{bc_waveica}
\alias{bc_waveica}
\title{WaveICA_2.0 batch correction}
\usage{
bc_waveica(omicsData, injection_cname, alpha = 0, cutoff = 0.1, K = 10)
}
\arguments{
\item{omicsData}{an object of the class 'pepData', 'proData', 'metabData',
'lipidData', or 'nmrData', usually created by \code{\link{as.pepData}},
\code{\link{as.proData}}, \code{\link{as.metabData}},
\code{\link{as.lipidData}}, or \code{\link{as.nmrData}}, respectively.}

\item{injection_cname}{character string giving the name of the column in omicsData$f_data
that contains the run order of the samples}

\item{alpha}{tradeoff value between the independence of samples and those of variables in ICA,
and should be between 0 and 1, defaults to 0}

\item{cutoff}{threshold of the variation explained by the injection order for independent components,
should be between 0 and 1, defaults to 0.1}

\item{K}{maximal component that ICA decomposes, defaults to 10}
}
\value{
Object of same class as omicsData that has been undergone
  WaveICA normalization
}
\description{
This function returns a pmart object that has been undergone WaveICA_2.0
 batch effect correction
}
\examples{
library(malbacR)
library(pmartR)
data("pmart_amide")
pmart_amide <- edata_transform(pmart_amide,"log2")
pmart_amide <- group_designation(pmart_amide,main_effects = "group",batch_id = "batch")
pmart_amide <- normalize_global(pmart_amide,subset_fn = "all",norm_fn = "median",
                               apply_norm = TRUE,backtransform = TRUE)
impObj <- imputation(omicsData = pmart_amide)
amide_imp <- apply_imputation(imputeData = impObj, omicsData = pmart_amide)
amide_wave <- bc_waveica(omicsData = amide_imp, injection_cname = "Injection_order",
                         alpha = 0, cutoff = 0.1, K = 10)

}
\author{
Damon Leach
}
