% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tiger.R
\name{compute_targetVal}
\alias{compute_targetVal}
\title{compute_targetVal}
\usage{
compute_targetVal(
  QC_num,
  sampleType,
  batchID = NULL,
  targetVal_method = c("mean", "median"),
  targetVal_batchWise = FALSE,
  targetVal_removeOutlier = !targetVal_batchWise,
  coerce_numeric = FALSE
)
}
\arguments{
\item{QC_num}{description}

\item{sampleType}{description}

\item{batchID}{description}

\item{targetVal_method}{description}

\item{targetVal_batchWise}{description}

\item{targetVal_removeOutlier}{description}

\item{coerce_numeric}{description}
}
\value{
return created folds
}
\description{
This function to compute target values
}
