% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fns.R
\name{get_batch_method}
\alias{get_batch_method}
\title{Fetch the Batch Effect Method Name}
\usage{
get_batch_method(omicsData)
}
\arguments{
\item{omicsData}{An object of class pepData, proData, metabData, lipidData,
or nmrData.}
}
\value{
A character vector specifying the batch correction method used (NULL if no batch correction applied)
}
\description{
Retrieves the name of the batch effect
}
