% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{imputation}
\alias{imputation}
\title{imputation}
\usage{
imputation(omicsData)
}
\arguments{
\item{omicsData}{an object of the class 'pepData', 'proData', 'metabData',
'lipidData', or 'nmrData', usually created by \code{\link{as.pepData}},
\code{\link{as.proData}}, \code{\link{as.metabData}},
\code{\link{as.lipidData}}, or \code{\link{as.nmrData}}, respectively.}
}
\value{
object of class imputeData and dataframe with the imputated e_data
}
\description{
This function returns a pmart filter that has imputed missing values by "em" method
}
\examples{
library(malbacR)
data("pmart_amide")
library(pmartR)
pmart_amide <- edata_transform(pmart_amide,"log2")
impFilt <- imputation(pmart_amide)

}
\author{
Evan Martin, Damon Leach
}
