% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tiger.R
\name{Internal.compute_cor}
\alias{Internal.compute_cor}
\title{Internal.compute_cor}
\usage{
Internal.compute_cor(
  train_num,
  test_num = NULL,
  selectVar_corType = c("pcor", "cor"),
  selectVar_corMethod = c("spearman", "pearson")
)
}
\value{
return correlation computed
}
\description{
This function is an internal function to compute correlation
}
