% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tiger.R
\name{Internal.run_ensemble}
\alias{Internal.run_ensemble}
\title{Internal.run_ensemble}
\usage{
Internal.run_ensemble(
  trainSet,
  testSet,
  mtry_percent = seq(0.2, 0.8, 0.2),
  nodesize_percent = seq(0.2, 0.8, 0.2),
  ...,
  return_base_res = FALSE
)
}
\value{
return ensemble learning
}
\description{
This function is an internal function to run ensemble learning
}
